+++ src/main/java/com/github/manevolent/jbot/command/response/DefaultBasicCommandListResponse.java
@@ -5,8 +5,8 @@ import com.github.manevolent.jbot.command.exception.CommandExecutionException;
 
 public class DefaultBasicCommandListResponse<T> extends CommandListResponse<T> {
     public DefaultBasicCommandListResponse(ChatSender sender,
-                                           int actualTotal,
-                                           int page,
+                                           long actualTotal,
+                                           long page,
                                            int elementsPerPage,
                                            ListAccessor<T> accessor,
                                            ListElementFormatter<T> responder) {
@@ -15,8 +15,8 @@ public class DefaultBasicCommandListResponse<T> extends CommandListResponse<T> {
 
     @Override
     public void send() throws CommandExecutionException {
-        int totalPages = (int) Math.ceil((double)getTotalElements() / (double)getElementsPerPage());
-        int elements = Math.min(getAccessor().size(), getElementsPerPage());
+        long totalPages = (long) Math.ceil((double)getTotalElements() / (double)getElementsPerPage());
+        long elements = Math.min(getAccessor().size(), getElementsPerPage());
 
         if (elements < 0) throw new CommandExecutionException("Invalid page (" + totalPages + " pages).");
         else if (elements == 0) throw new CommandExecutionException("No results found.");
