+++ src/main/java/com/github/manevolent/jbot/command/response/DefaultRichCommandListResponse.java
@@ -10,7 +10,7 @@ import java.util.Collection;
 import java.util.LinkedList;
 
 public class DefaultRichCommandListResponse<T> extends CommandListResponse<T> {
-    public DefaultRichCommandListResponse(ChatSender sender, int actualTotal, int page,
+    public DefaultRichCommandListResponse(ChatSender sender, long actualTotal, long page,
                                           int elementsPerPage, ListAccessor<T> accessor,
                                           ListElementFormatter<T> responder) {
         super(sender, actualTotal, page, elementsPerPage, accessor, responder);
@@ -19,7 +19,7 @@ public class DefaultRichCommandListResponse<T> extends CommandListResponse<T> {
     @Override
     public void send() throws CommandExecutionException {
         int totalPages = (int) Math.ceil((double)getTotalElements() / (double)getElementsPerPage());
-        int elements = Math.min(getAccessor().size(), getElementsPerPage());
+        long elements = Math.min(getAccessor().size(), getElementsPerPage());
 
         if (elements < 0) throw new CommandExecutionException("Invalid page (" + totalPages + " pages).");
         else if (elements == 0) throw new CommandExecutionException("No results found.");
